corrected unsigned long long int

+++ LinearDS/wormholes/wormholes.cpp
@@ -3,16 +3,15 @@
 # include <bits/stdc++.h>
 # include <cstdio>
 
-int takeTest (std::pair<int, int> testTimes[], int wormV[], int wormW[], int N, int X, int Y)
+unsigned long long int takeTest (std::pair<int, int> testTimes[], int wormV[], int wormW[], int N, int X, int Y)
 {
 	 // initialising bestTime to the largest possible time
-	 int bestTime = wormW[Y-1] - wormV[0] + 1;
+	 unsigned long long int bestTime = wormW[Y-1] - wormV[0] + 1;
 	 for (int i=0; i<N; i++)
 	 {
 			int entry= 0;
 			int exit = Y-1;
-			// 	int flag1 = 0;
-			// 	int flag2 = 0;
+
 			if ((testTimes[i].second - testTimes[i].first + 1) < bestTime)
 			{
 				 for (int j=X-1; j>0; j--)
@@ -31,7 +30,7 @@ int takeTest (std::pair<int, int> testTimes[], int wormV[], int wormW[], int N, 
 							 break;
 						}
 				 }
-				 int thisTime = wormW[exit] - wormV[entry] + 1;
+				 unsigned long long int thisTime = wormW[exit] - wormV[entry] + 1;
 				 if (thisTime < bestTime)
 				 {
 						bestTime = thisTime;
